Mapping_Tagging/requirements.txt
# Core dependencies
fastapi>=0.103.0
uvicorn>=0.23.2
pydantic>=2.3.0
python-dotenv>=1.0.0
logfire>=0.8.0

# AI/ML dependencies
openai>=1.0.0
pydantic-ai>=0.1.0

# HTTP and API utilities
httpx>=0.25.0
requests>=2.31.0
aiohttp>=3.8.5

# Typing and utilities
typing-extensions>=4.7.1
python-multipart>=0.0.6
jinja2>=3.1.2

# Data processing
pandas>=2.1.0

# Development tools
pytest>=7.4.0
pytest-asyncio>=0.21.1
black>=23.7.0
isort>=5.12.0